fixed bounding box calculations that used int values when floats are possible (now correctly using floats)

+++ cvdata/convert.py
@@ -379,12 +379,12 @@ def kitti_to_darknet(
                 else:
                     label_index = len(label_indices)
                     label_indices[label] = label_index
-                box_width_pixels = int(parts[6]) - int(parts[4]) + 1
-                box_height_pixels = int(parts[7]) - int(parts[5]) + 1
+                box_width_pixels = float(parts[6]) - float(parts[4]) + 1
+                box_height_pixels = float(parts[7]) - float(parts[5]) + 1
                 darknet_bbox = {
                     "label_index": label_index,
-                    "center_x": ((box_width_pixels / 2)  + int(parts[4])) / width,
-                    "center_y": ((box_height_pixels / 2)  + int(parts[5])) / height,
+                    "center_x": ((box_width_pixels / 2)  + float(parts[4])) / width,
+                    "center_y": ((box_height_pixels / 2)  + float(parts[5])) / height,
                     "box_width": box_width_pixels / width,
                     "box_height": box_height_pixels / height,
                 }
